requirements.txt
aiohttp==3.13.5
python-dotenv==1.2.2
